refactor: split video uploads out of ProductImageUploads and fix missing videos translation key
