+++ backend/gateway/src/locales/en/product.json
@@ -112,6 +112,7 @@
     "attachment": "Attachment",
     "featured-image": "Featured image",
     "secondary-images": "Secondary images",
+    "videos": "Videos",
     "product-variants": "Product Variants",
     "uom-description": "Choose the unit of measure for the product",
     "notFound": "Product not found",
