+++ backend/gateway/src/locales/mn/product.json
@@ -110,6 +110,7 @@
     "attachment": "Хавсралт",
     "featured-image": "Үндсэн зураг",
     "secondary-images": "Нэмэлт зургууд",
+    "videos": "Видео",
     "product-variants": "Бүтээгдэхүүний хувилбарууд",
     "uom-description": "Бүтээгдэхүүний хэмжих нэгжийг сонгоно уу",
     "notFound": "Бүтээгдэхүүн олдсонгүй",
